refactor(liblili2json): move clean_word to utils.py
